dumpling: Fix inaccurate notification LED colors
The red LED on dumpling appears to be very weak compared to the green
and blue LEDs (possibly due to poor placement?), so color mixes involving
red come out inaccurate (e.g. "white" has a very strong cyan hue).

In order to somewhat alleviate the color imbalance, reduce the brightness
of the green and blue LEDs by 50% when a color mix involving red is
requested. This reduction still isn't enough to fix the color inaccuracy
(90/255 was empirically found to be the best fraction of the blue/green
LED brightness to use), but reducing the brightness of the blue/green
LEDs any further would make the notification LED appear too dim.